+++ src/main/kotlin/com/ostaragame/systems/economy/engine/TradeGood.kt
@@ -3,4 +3,4 @@ package com.ostaragame.systems.economy.engine
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class TradeGood(val name: String, val type: TradeLibrary.TradeGoodType, var cost: Float, val volume: Float)
+data class TradeGood(val name: String, val type: TradeLibrary.TradeGoodType, var cost: Int, val volume: Int)
